requirements.txt
torch
torchaudio
hydra-core
pytorch-lightning
librosa
